+++ tests/RequestHandlerTest.php
@@ -6,7 +6,7 @@ namespace DoppioGancio\MockedClient\Tests;
 
 use DoppioGancio\MockedClient\Exception\RouteNotFound;
 use DoppioGancio\MockedClient\RequestHandler;
-use DoppioGancio\MockedClient\Route\RouteBuilder;
+use DoppioGancio\MockedClient\Route\Route;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use Http\Discovery\Psr17FactoryDiscovery;
@@ -22,17 +22,8 @@ class RequestHandlerTest extends TestCase
             new NullLogger(),
         );
 
-        $routeBuilder = new RouteBuilder(
-            Psr17FactoryDiscovery::findResponseFactory(),
-            Psr17FactoryDiscovery::findStreamFactory(),
-        );
-
         $requestHandler->addRoute(
-            $routeBuilder
-                ->withMethod('GET')
-                ->withPath('/country/IT')
-                ->withResponse(new Response(200, [], '{"code":"IT"}'))
-                ->build(),
+            new Route('GET', '/country/IT', static fn () => new Response(200, [], '{"code":"IT"}')),
         );
 
         $response = $requestHandler->handle(new Request('GET', 'http://www.any.com/country/IT'));